qna-backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-genai
requests
aiohttp
sqlalchemy
databases
asyncpg
psycopg2-binary
